made movie list a component, added list of movies for each theatre, sorted movies in movie slice
