+++ DrCinema/src/slices/moviesSlice.js
@@ -30,7 +30,9 @@ const moviesSlice = createSlice({
         });
         builder.addCase(fetchMovies.fulfilled, (state, action) => {
             state.isLoading = false;
-            state.data = action.payload;
+            const data = action.payload;
+            const sortedData = data ? [...data].sort((movieA, movieB) => movieA.title.localeCompare(movieB.title, 'is', { sensitivity: "base" })) : null;
+            state.data = sortedData;
         });
         builder.addCase(fetchMovies.rejected, (state, action) => {
             state.isLoading = false;
